move util functions and associated unit tests to the appropriate slice (from old DagNodeSlice to TreeSlice)
